+++ extensions/raycast-wallpaper/CHANGELOG.md
@@ -4,3 +4,4 @@
 - Release of Raycast Wallpaper version 1.0
 - **Raycast Wallpaper**: Get and set Raycast wallpapers to desktop.
 - **Random Wallpaper**: Set random Raycast wallpapers to desktop.
+- **Capture Raycast Metadata**: Capture a screenshot for Raycast Extension's metadata.